# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StreamGuard API - Batch IPTV stream health checker and playlist manager
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all check jobs
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Start a new check job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Job created and queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job status and live stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job and its results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /jobs/{id}/pause:
    post:
      operationId: pauseJob
      tags: [jobs]
      summary: Pause a running job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}/resume:
    post:
      operationId: resumeJob
      tags: [jobs]
      summary: Resume a paused job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}/cancel:
    post:
      operationId: cancelJob
      tags: [jobs]
      summary: Cancel a running or paused job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}/results:
    get:
      operationId: getJobResults
      tags: [jobs]
      summary: Get check results for a job with filtering/sorting/pagination
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortDir
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultPage"
  /jobs/{id}/summary:
    get:
      operationId: getJobSummary
      tags: [jobs]
      summary: Get live stats summary for dashboard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobSummary"
  /jobs/{id}/categories:
    get:
      operationId: getJobCategories
      tags: [jobs]
      summary: Get category breakdown with counts for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category list with counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryCount"
  /jobs/{id}/probe:
    post:
      operationId: probeChannels
      tags: [jobs]
      summary: Trigger Stage 3 deep ffprobe on selected result IDs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProbeRequest"
      responses:
        "202":
          description: Probe queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /jobs/{id}/export:
    get:
      operationId: exportJob
      tags: [jobs]
      summary: Export filtered results as M3U, CSV, or JSON
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: format
          in: query
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            description: Comma-separated status filter e.g. "live,suspicious"
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: File download
          content:
            text/plain:
              schema:
                type: string
            text/csv:
              schema:
                type: string
            application/json:
              schema:
                type: string
  /jobs/{id}/diff/{compareId}:
    get:
      operationId: diffJobs
      tags: [jobs]
      summary: Compare two jobs of the same playlist - show what changed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: compareId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Diff between two job runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDiff"

  # ── Results ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ApiMessage:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Playlist schemas ────────────────────────────────────────────────────────
...
    JobSettings:
      type: object
      properties:
        concurrency:
          type: integer
          default: 30
        timeoutMs:
          type: integer
          default: 8000
        retryCount:
          type: integer
          default: 3
        autoProbe:
          type: boolean
          default: false
        userAgents:
          type: array
          items:
            type: string
      required: []
    Job:
      type: object
      properties:
        id:
          type: integer
        playlistId:
          type: integer
        playlistName:
          type: string
        status:
          type: string
          enum: [queued, running, paused, completed, cancelled, failed]
        settings:
          $ref: "#/components/schemas/JobSettings"
        total:
          type: integer
        checked:
          type: integer
        live:
          type: integer
        dead:
          type: integer
        geoblocked:
          type: integer
        suspicious:
          type: integer
        pending:
          type: integer
        etaSeconds:
          type: ["integer", "null"]
        avgCheckMs:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
      required: [id, playlistId, playlistName, status, settings, total, checked, live, dead, geoblocked, suspicious, pending, createdAt]
    JobInput:
      type: object
      properties:
        playlistId:
          type: integer
        settings:
          $ref: "#/components/schemas/JobSettings"
      required: [playlistId]
    JobSummary:
      type: object
      properties:
        jobId:
          type: integer
        status:
          type: string
        total:
          type: integer
        checked:
          type: integer
        live:
          type: integer
        dead:
          type: integer
        geoblocked:
          type: integer
        suspicious:
          type: integer
        pending:
          type: integer
        livePercent:
          type: number
        deadPercent:
          type: number
        avgCheckMs:
          type: ["number", "null"]
        etaSeconds:
          type: ["integer", "null"]
        topCategories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
        progressPercent:
          type: number
      required: [jobId, status, total, checked, live, dead, geoblocked, suspicious, pending, livePercent, deadPercent, progressPercent, topCategories]
    CategoryCount:
      type: object
      properties:
        category:
          type: string
        total:
          type: integer
        live:
          type: integer
        dead:
          type: integer
        geoblocked:
          type: integer
        suspicious:
          type: integer
        pending:
          type: integer
      required: [category, total, live, dead, geoblocked, suspicious, pending]
    JobDiff:
      type: object
      properties:
        jobId:
          type: integer
        compareJobId:
          type: integer
        newlyDead:
          type: array
          items:
            $ref: "#/components/schemas/ChannelResult"
        newlyLive:
          type: array
          items:
            $ref: "#/components/schemas/ChannelResult"
        unchanged:
          type: integer
        summary:
          type: string
      required: [jobId, compareJobId, newlyDead, newlyLive, unchanged, summary]

    # ── Result schemas ──────────────────────────────────────────────────────────
    ProbeData:
      type: object
      properties:
        videoCodec:
          type: ["string", "null"]
        audioCodec:
          type: ["string", "null"]
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
        framerate:
          type: ["number", "null"]
        bitrate:
          type: ["integer", "null"]
        container:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        mislabeled:
          type: boolean
        mislabelReason:
          type: ["string", "null"]
      required: []
    ChannelResult:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        channelId:
          type: integer
        tvgName:
          type: ["string", "null"]
        tvgLogo:
          type: ["string", "null"]
        url:
          type: string
        category:
          type: ["string", "null"]
        status:
          type: string
          enum: [live, dead, geoblocked, suspicious, pending]
        httpStatus:
          type: ["integer", "null"]
        responseTimeMs:
          type: ["integer", "null"]
        redirectCount:
          type: ["integer", "null"]
        tlsValid:
          type: ["boolean", "null"]
        mimeType:
          type: ["string", "null"]
        manifestValid:
          type: ["boolean", "null"]
        failureReason:
          type: ["string", "null"]
        probeData:
          $ref: "#/components/schemas/ProbeData"
        checkedAt:
          type: ["string", "null"]
          format: date-time
      required: [id, jobId, channelId, url, status]
    ResultPage:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/ChannelResult"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [results, total, page, limit]
...
    ProbeRequest:
      type: object
      properties:
        resultIds:
          type: array
          items:
            type: integer
      required: [resultIds]

    # ── Settings schemas ────────────────────────────────────────────────────────
...